+++ tests/test_text_data.py
@@ -40,7 +40,7 @@ def test_from_csv():
         filepath = os.path.join(path, filename+'.csv')
         try:
             text_csv_file(filepath, n_labels=n_labels)
-            data = TextClasDataBunch.from_csv(path, train=filename, valid=filename, n_labels=n_labels)
+            data = TextClasDataBunch.from_csv(path, train=filename, valid=filename, test=filename, n_labels=n_labels)
             assert len(data.classes) == 2
             assert set(data.classes) == set([True, False])
             if n_labels > 1: assert len(data.labels[0]) == n_labels
@@ -53,9 +53,18 @@ def test_from_df():
         os.makedirs(path)
         try:
             df = text_df(n_labels=n_labels)
-            data = TextClasDataBunch.from_df(path, train_df=df, valid_df=df, label_cols=list(range(n_labels)), txt_cols=["text"])
+            data = TextClasDataBunch.from_df(path, train_df=df, valid_df=df, test_df=df, label_cols=list(range(n_labels)), txt_cols=["text"])
             assert len(data.classes) == 2
             assert set(data.classes) == set([True, False])
             if n_labels > 1: assert len(data.labels[0]) == n_labels
         finally:
             shutil.rmtree(path)
+            
+def test_collate():
+    path = untar_data(URLs.IMDB_SAMPLE)
+    dft = pd.read_csv(path/'train.csv', header=None)
+    dfv = pd.read_csv(path/'valid.csv', header=None)
+    data = TextClasDataBunch.from_df(path, dft, dfv, bs=20)
+    x,y = next(iter(data.train_dl))
+    assert x.size(0) == 1519 and x.size(1) == 10
+    assert x[0,-1] == 1